Add functions to check the presence of configuration variables
for both Viber and Telegram bots in the configuration file.
